pyproject.toml
[project]
name = "cookiefyapi"
version = "0.1.0"
description = "Cookiefy API project to handle the application services"
authors = [
    {name = "Ferran Colome",email = "ferrancolomsanz@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.14.1,<2.0.0)",
    "pydantic-settings (>=2.8.0,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
mypy = "^1.15.0"
pytest = "^8.3.4"
httpx = "^0.28.1"

